requirements/tests.txt
django-discover-runner